requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-cors==0.8.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
backoff==2.2.1
beartype==0.15.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.2
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.2
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
dataclasses-json==0.6.7
decorator==5.2.1
deltalake==0.17.4
Deprecated==1.2.18
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
docling==2.30.0
docling-core==2.26.3
docling-ibm-models==3.4.1
docling-parse==4.0.1
easyocr==1.7.2
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fs==2.4.16
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.29.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h3==4.2.2
html5lib==1.1
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.14
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.30
latex2mathml==3.77.0
lazy_loader==0.4
linkify-it-py==2.0.3
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.73.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
panel==1.6.2
param==2.2.0
parso==0.8.4
pathway==0.21.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
python-pptx==1.0.2
python-sat==1.8.dev16
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
rtree==1.4.0
s3transfer==0.10.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semchunk==2.2.2
sentence-transformers==4.0.2
setuptools==78.1.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlglot==10.6.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unstructured==0.16.11
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
wrapt==1.17.2
XlsxWriter==3.2.2
xyzservices==2025.1.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
